+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Add enum constraints to unit fields, and replace TODOs.
 - Check that directory schemas exist.
 ### Changed
+- Simplified directory validation.
 - mass -> mz
 - bulkrnaseq QA is just RIN.
 - Add bytes to IMC.